+++ src/main/java/com/aek56/microservice/auth/model/security/AuthUser.java
@@ -40,7 +40,7 @@ public class AuthUser extends AbstractAuthUser {
     /**
      * 机构Id
      */
-    private Integer tenantId;
+    private Long tenantId;
     
     /**
      * 密码
@@ -185,12 +185,12 @@ public class AuthUser extends AbstractAuthUser {
         return enabled;
     }
     
-    public Integer getTenantId()
+    public Long getTenantId()
     {
         return tenantId;
     }
     
-    public void setTenantId(Integer tenantId)
+    public void setTenantId(Long tenantId)
     {
         this.tenantId = tenantId;
     }
